Add cemu existence checks

+++ cemugraphicspacks-uninstall.ps1
@@ -1,5 +1,10 @@
-Get-ChildItem $dir -Recurse -Directory |
-ForEach-Object {
-	write-output "Unlinking $($_.Name)"
-	rm -r -Force "$dir\..\..\cemu\current\graphicPacks\$($_.Name)"
+$cemuExists = (New-Object System.IO.FileInfo("$dir\..\..\cemu\current")).Exists
+if ($cemuExists) {
+	Get-ChildItem $dir -Recurse -Directory |
+	ForEach-Object {
+		write-output "Unlinking $($_.Name)"
+		rm -r -Force "$dir\..\..\cemu\current\graphicPacks\$($_.Name)"
+	}
+} else {
+	write-output "Cannot find cemu folder. If it has already been uninstalled, this is not a problem."
 }
